+++ commands/vcf/main.py
@@ -3,6 +3,8 @@ import logging
 from pathlib import Path
 from typing import List, Optional
 
+from commands.eval.classes.run_settings import RunSettings
+from commands.eval.classes.score_paths import ScorePaths
 from shared.vcf.main_functions import variant_comparisons
 
 logger = logging.getLogger(__name__)
@@ -19,15 +21,20 @@ def main(
     run_id1: Optional[str],
     run_id2: Optional[str],
     results: Optional[Path],
-    annotations: List[str],
+    extra_annot_keys: List[str],
+    output_all_variants: bool,
 ):
-    show_line_numbers = True
+    run_settings = RunSettings(
+        score_threshold=score_threshold,
+        max_display=max_display,
+        max_checked_annots=max_checked_annots,
+        show_line_numbers=True,
+        extra_annot_keys=extra_annot_keys,
+    )
 
     if results is not None:
         if not results.exists():
             results.mkdir(parents=True)
-
-    out_path_presence = results / "presence.txt" if results else None
 
     if run_id1 is None:
         run_id1 = str(vcf1).split("/")[-1].split(".")[0]
@@ -39,8 +46,8 @@ def main(
             run_id2 = run_id2 + "_2"
         logger.info(f"# --run_id2 not set, assigned: {run_id2}")
 
-    out_path_score_above_thres = results / "above_thres.txt" if results else None
-    out_path_score_all = results / "score_all.txt" if results else None
+    label = "sv" if is_sv else "snv"
+    score_paths = ScorePaths(label, results, score_threshold, output_all_variants)
 
     do_score_check = True
     do_annot_check = True
@@ -51,32 +58,30 @@ def main(
         vcf1,
         vcf2,
         is_sv,
-        score_threshold,
-        max_display,
-        max_checked_annots,
-        out_path_presence,
-        out_path_score_above_thres,
-        out_path_score_all,
+        run_settings,
+        score_paths,
         do_score_check,
         do_annot_check,
-        show_line_numbers,
-        annotations,
     )
 
 
 def main_wrapper(args: argparse.Namespace):
 
+    if args.silent:
+        logger.setLevel(logging.WARNING)
+
     main(
         args.vcf1,
         args.vcf2,
         args.is_sv,
-        args.score_threshold,
-        args.max_checked_annots,
         args.max_display,
+        args.max_checked_annots,
+        args.score_threshold,
         args.id1,
         args.id2,
         args.results if args.results is not None else None,
         args.annotations.split(",") if args.annotations is not None else [],
+        args.all_variants,
     )
 
 
@@ -120,6 +125,14 @@ def add_arguments(parser: argparse.ArgumentParser):
     parser.add_argument(
         "--annotations",
         help="Comma separated additional annotations to retain in output",
+    )
+    parser.add_argument(
+        "--silent", action="store_true", help="Run silently, produce only output files"
+    )
+    parser.add_argument(
+        "--all_variants",
+        action="store_true",
+        help="Write a comparison file including non-differing variants",
     )