+++ tests/test_eval.py
@@ -6,6 +6,8 @@ from typing import List
 import pytest
 from pytest import LogCaptureFixture
 
+from commands.eval.classes.run_object import RunObject
+from commands.eval.classes.run_settings import RunSettings
 from commands.eval.main import main
 
 
@@ -85,52 +87,45 @@ def test_eval_main(
     caplog: LogCaptureFixture,
     mock_results: tuple[Path, Path, Path],
 ):
+    """
+    Given a valid dataset, check that the command runs without errors,
+    outputs expected files, and do some sanity checking on the score comparisons
+    """
     results1, results2, outdir = mock_results
 
+    run_object = RunObject("r1", "r2", results1, results2)
+    run_settings = RunSettings(score_threshold=17)
+
     with caplog.at_level(logging.INFO):
-        main(
-            "r1",
-            "r2",
-            results1,
-            results2,
-            None,
-            None,
-            17,
-            15,
-            outdir,
-            False,
-            1000,
-            False,
-            [],
-        )
+        main(run_object, run_settings, None, None, outdir)
 
     assert len(caplog.records) > 0, "No logs were captured"
     assert not any(record.levelname == "ERROR" for record in caplog.records)
 
-    expected = [
+    expected_files: List[str] = [
         "check_sample_files.txt",
         "scored_snv_presence.txt",
-        "scored_snv_score_thres_17.txt",
-        "scored_snv_score_all.txt",
+        "scored_snv_above_thres_17.txt",
+        "scored_snv_all.txt",
         "scored_sv_presence.txt",
-        "scored_sv_score_thres_17.txt",
-        "scored_sv_score.txt",
+        "scored_sv_above_thres_17.txt",
+        "scored_sv_all.txt",
         "yaml_diff.txt",
     ]
 
-    for fname in expected:
+    for fname in expected_files:
         assert (outdir / fname).exists(), f"Expected file {fname} does not exist"
 
     # Verify that differences were detected and written to the output files
     snv_score_thres = (
-        (outdir / "scored_snv_score_thres_17.txt").read_text().splitlines()
+        (outdir / "scored_snv_above_thres_17.txt").read_text().splitlines()
     )
     assert any(
         "G/T" in line for line in snv_score_thres
     ), "Expected SNV difference missing"
 
-    snv_score_all = (outdir / "scored_snv_score_all.txt").read_text().splitlines()
+    snv_score_all = (outdir / "scored_snv_all.txt").read_text().splitlines()
     assert len(snv_score_all) == 3
 
-    sv_score = (outdir / "scored_sv_score.txt").read_text().splitlines()
+    sv_score = (outdir / "scored_sv_all.txt").read_text().splitlines()
     assert any("<DEL>" in line for line in sv_score)